Added redis get_values()
